requirements.txt
attrs==19.1.0
black==0.0
chardet==2.3.0
Click==7.0
cryptography==1.7.1
Flask==1.0.3
idna==2.2
itsdangerous==1.1.0
Jinja2==2.10.1
keyring==10.1
keyrings.alt==1.3
MarkupSafe==1.1.1
pyasn1==0.1.9
pycrypto==2.6.1
pygobject==3.22.0
PyJWT==1.7.1
PySocks==1.7.0
python-apt==1.1.0b5
pytz==2019.1
pyxdg==0.25
requests==2.12.4
RPi.GPIO==0.6.5
SecretStorage==2.3.1
six==1.12.0
ssh-import-id==5.6
twilio==6.27.1
urllib3==1.19.1
Werkzeug==0.15.4
